refactor: adopt feature folders, typed routes, and sync pages
Moves the blog domain into features/post/ (queries, blog index, posts
list, series nav), shrinks app/blog/page.tsx to composition only, and
switches pages to params.then()/searchParams.then() so page chrome
stays outside the data path. Enables typedRoutes with Route-typed
link props. Rendered HTML is unchanged.
