+++ app/blog/[slug]/page.tsx
@@ -4,13 +4,13 @@ import Link from "next/link"
 import { notFound } from "next/navigation"
 
 import { siteConfig } from "@/lib/config"
+import { calculateReadingTime } from "@/lib/reading-time"
 import {
   formatPostDate,
   getAllPosts,
   getPostBySlug,
   getPostTags,
-} from "@/lib/posts"
-import { calculateReadingTime } from "@/lib/reading-time"
+} from "@/features/post/post-queries"
 import { MDXContent } from "@/components/mdx-components"
 import MdxLayout from "@/components/mdx-layout"
 
@@ -19,9 +19,9 @@ export const dynamicParams = false
 export const generateStaticParams = async () =>
   getAllPosts(false).map((post) => ({ slug: post._meta.path }))
 
-export const generateMetadata = async (props: {
-  params: Promise<{ slug: string }>
-}): Promise<Metadata> => {
+export const generateMetadata = async (
+  props: PageProps<"/blog/[slug]">
+): Promise<Metadata> => {
   const params = await props.params
   const post = getPostBySlug(params.slug)
   if (!post) notFound()
@@ -64,59 +64,58 @@ export const generateMetadata = async (props: {
   }
 }
 
-const PostLayout = async (props: { params: Promise<{ slug: string }> }) => {
-  const params = await props.params
-  const post = getPostBySlug(params.slug)
-  if (!post) notFound()
+export default function PostPage({ params }: PageProps<"/blog/[slug]">) {
+  return params.then(({ slug }) => {
+    const post = getPostBySlug(slug)
+    if (!post) notFound()
 
-  const readingTime = calculateReadingTime(post.content)
-  const tags = getPostTags(post.tag)
+    const readingTime = calculateReadingTime(post.content)
+    const tags = getPostTags(post.tag)
 
-  return (
-    <article className="py-8 md:py-16">
-      <Link
-        href="/blog"
-        className="focus-visible:ring-offset-lightGray dark:focus-visible:ring-offset-dark mb-8 inline-flex min-h-touch items-center rounded-lg text-sm text-gray-600 transition-colors hover:text-purple-700 focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-purple-500/70 focus-visible:ring-offset-2 dark:text-gray-400 dark:hover:text-purple-300 md:mb-10"
-      >
-        Back to articles
-      </Link>
-
-      <header className="mb-8 max-w-3xl space-y-4 md:mb-12 md:space-y-5">
-        <div className="flex flex-wrap gap-2">
-          {tags.map((tag) => (
-            <span
-              key={tag}
-              className="rounded-full border border-purple-500/30 px-3 py-1 text-xs text-purple-700 dark:text-purple-300"
-            >
-              {tag}
-            </span>
-          ))}
-        </div>
-        <ViewTransition
-          name={`post-title-${params.slug}`}
-          share="text-morph"
-          default="none"
+    return (
+      <article className="py-8 md:py-16">
+        <Link
+          href="/blog"
+          className="focus-visible:ring-offset-lightGray dark:focus-visible:ring-offset-dark mb-8 inline-flex min-h-touch items-center rounded-lg text-sm text-gray-600 transition-colors hover:text-purple-700 focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-purple-500/70 focus-visible:ring-offset-2 dark:text-gray-400 dark:hover:text-purple-300 md:mb-10"
         >
-          <h1 className="text-wrap font-heading text-[2rem] font-bold leading-tight tracking-tight text-gray-900 dark:text-gray-50 md:text-6xl">
-            {post.title}
-          </h1>
-        </ViewTransition>
-        <p className="text-base leading-7 text-gray-700 dark:text-gray-300 md:text-lg md:leading-relaxed">
-          {post.summary}
-        </p>
-        <div className="flex flex-wrap gap-x-5 gap-y-2 text-sm text-gray-500 dark:text-gray-400">
-          <time dateTime={post.date}>
-            Published {formatPostDate(post.date)}
-          </time>
-          <span>{readingTime} min read</span>
-        </div>
-      </header>
+          Back to articles
+        </Link>
 
-      <MdxLayout>
-        <MDXContent source={post.content} />
-      </MdxLayout>
-    </article>
-  )
+        <header className="mb-8 max-w-3xl space-y-4 md:mb-12 md:space-y-5">
+          <div className="flex flex-wrap gap-2">
+            {tags.map((tag) => (
+              <span
+                key={tag}
+                className="rounded-full border border-purple-500/30 px-3 py-1 text-xs text-purple-700 dark:text-purple-300"
+              >
+                {tag}
+              </span>
+            ))}
+          </div>
+          <ViewTransition
+            name={`post-title-${slug}`}
+            share="text-morph"
+            default="none"
+          >
+            <h1 className="text-wrap font-heading text-[2rem] font-bold leading-tight tracking-tight text-gray-900 dark:text-gray-50 md:text-6xl">
+              {post.title}
+            </h1>
+          </ViewTransition>
+          <p className="text-base leading-7 text-gray-700 dark:text-gray-300 md:text-lg md:leading-relaxed">
+            {post.summary}
+          </p>
+          <div className="flex flex-wrap gap-x-5 gap-y-2 text-sm text-gray-500 dark:text-gray-400">
+            <time dateTime={post.date}>
+              Published {formatPostDate(post.date)}
+            </time>
+            <span>{readingTime} min read</span>
+          </div>
+        </header>
+
+        <MdxLayout>
+          <MDXContent source={post.content} />
+        </MdxLayout>
+      </article>
+    )
+  })
 }
-
-export default PostLayout
