+++ components/mdx-components.tsx
@@ -1,3 +1,4 @@
+import type { Route } from "next"
 import Image from "next/image"
 import Link from "next/link"
 import type { MDXComponents } from "mdx/types"
@@ -7,13 +8,13 @@ import rehypePrettyCode from "rehype-pretty-code"
 import remarkGfm from "remark-gfm"
 
 import { cn } from "@/lib/utils"
+import { SeriesNav } from "@/features/post/components/series-nav"
 import { Banner } from "@/components/banner"
 import { Callout } from "@/components/callout"
 import { Code } from "@/components/code"
 import { MdxCard, MdxDisabledCard, MdxLinkCard } from "@/components/mdx-card"
 import { MdxImage } from "@/components/mdx-image"
 import { PluginCard } from "@/components/plugin-card"
-import { SeriesNav } from "@/components/series-nav"
 import { VideoEmbed } from "@/components/video-embed"
 
 const mdxOptions: MDXRemoteProps["options"] = {
@@ -103,7 +104,7 @@ const components = {
 
     if (isInternal) {
       return (
-        <Link href={href!} className={cn(className)} {...props}>
+        <Link href={href as Route} className={cn(className)} {...props}>
           {children}
         </Link>
       )
