+++ next.config.mjs
@@ -3,6 +3,7 @@ import createMDX from "@next/mdx"
 /** @type {import('next').NextConfig} */
 const nextConfig = {
   reactStrictMode: true,
+  typedRoutes: true,
   experimental: {
     viewTransition: true,
   },
